Phuongtd login

+++ coffee-store/src/main/java/com/coffee/repository/account/IUserRoleRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface IUserRoleRepository extends JpaRepository<UserRole, Long> {
+public interface IUserRoleRepository extends JpaRepository<UserRole, Integer> {
 
     @Query(value = " select ur.id, ur.is_deleted, ur.role_id, ur.user_id from user_role ur where user_id = :#{#appUser.id} ", nativeQuery = true)
     List<UserRole> findAllByAppUser(@Param("appUser") AppUser appUser);
